+++ lib/pbkdf2/tools.ex
@@ -6,11 +6,12 @@ defmodule Pbkdf2.Tools do
   @allowed_chars 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
 
   def get_random_string(len, allowed_chars \\ @allowed_chars) do
+    :crypto.rand_seed()
     high = length(allowed_chars)
 
     Enum.to_list(1..len)
     |> Enum.reduce([], fn _x, acc ->
-      [Enum.at(allowed_chars, :crypto.rand_uniform(0, high))] ++ acc
+      [Enum.at(allowed_chars, :rand.uniform(high) - 1)] ++ acc
     end)
     |> to_string()
   end